renaming __init__.py to containers.py
